requirements.txt
Django==1.6.5
django-configurations==0.8
django-storages==1.1.8
Collectfast==0.2.1

boto==2.31.1
Fabric==1.9.0